ziyi: overlay: Refine display cutout region
By inspecting screen_round_corner_top.png from stock
product/overlay/DevicesOverlay.apk, we can observe that the cutout has
a radius of 37px and begins 25px below the top border. The cutout on
this device is not round but has a rectangle between the half circles
which is 168px wide. The rectangle approximation of the cutout should
have the same amount of padding on top of the cutout as below according
to the pixel device trees ("This configuartion effects status bar's
height actually. Set its height as double y position of
@dimen/config_mainBuiltInDisplayCutout to make cutout center vertically"
[1]). Additionally, add 8px padding at the left and right side of the
cutout to avoid content being too close to the cutout. Below is a sketch
of the cutout drawn within the rectangle approximation.
___________________
|                 |
|    _________    |  ___25px
|  /           \  |
| |             | |
|  \           /  |
|    ‾‾‾‾‾‾‾‾‾    |  ‾‾‾99px
|                 |
‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾  ‾‾124px
  |  |       |  | |
  8px|       213px|
     |          | |
     45px       250px
                  |
                  258px

[1]: https://android.googlesource.com/device/google/caimito/+/refs/heads/main/caiman/overlay/frameworks/base/core/res/res/values/config.xml#197

Change-Id: I117a347bbaddc42185b31127733fa7ee5078d147